feat(security): extract realm role converter and add context filter to enhance jwt handling and request processing
